==> 2. element.ipynb <==
# element: LogisticRegression

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
train_video_features= np.load(r"C:\Users\indra\Desktop\DSG PROJECT\video_features_noiseless_grayscale_resnet101.npy")

# the train csv file is loaded
df= pd.read_csv(r'C:\Users\indra\Desktop\DSG PROJECT\BH25\Training_Data\train.csv')

# since the motions are continuous value and the computer understands only number, the motions are label encoded before training
le = LabelEncoder()
df["element"] = le.fit_transform(df["element"])


df

from sklearn.model_selection import train_test_split
label= df["element"] # label is defined for y value of the model
X_train, X_val, y_train, y_val = train_test_split(train_video_features, label, test_size=0.2, random_state=42)


#logistic regression is used for this model

from sklearn.linear_model import LogisticRegression
model = LogisticRegression(max_iter=10000)
model.fit(train_video_features, label)

model.score(X_val, y_val) # checked the eval accuracy on seen data  

import joblib
joblib.dump(model, './element_model.pkl')
# the model is saved for further use



joblib.dump(le, 'label_encoder_element.pkl')
# the label encodes are used for further use


le.classes_
