==> 4. power.ipynb <==
import pandas as pd
df= pd.read_csv(r"C:\Users\indra\Desktop\DSG PROJECT\BH25\Training_Data\train.csv") # the train csv is loaded

df["power"].unique() # checked the unique powers available of which many are noises

data = {
    'power': ['rot', 'grin', 'geld', 'lila', 'blau', 'btau', 'rot', 'helb',
               'rod', 'grin', 'got', 'not', 'lila', 'trun', 'gela', 'gelt',
               'baau', 'liga', 'gelb', 'plau', 'brun', 'rot', 'lili', 'slau',
               'lisa', 'blru', 'grus', 'lula', 'gele', 'gaun']
}

# Count the occurrences of each unique element
value_counts = df['power'].value_counts()

value_counts # this was done to identify the outliers or noise


df = df[df['power'].isin(['rot', 'blau', 'lila', 'grun', 'gelb'])] # the rows with noise were droped and only 5 types of powers are kept for training

df

import cv2
import pandas as pd


def extract_brightest_color(video_path):
    cap = cv2.VideoCapture(video_path) # cv is to read the frames of the videos
    success, frame = cap.read()  # Reads the frames

    if success: # checks if frames are available
        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB) # converts the frame to rgb 
        brightest_pixel = frame_rgb.reshape(-1, 3).max(axis=0)  # gets RGB values of the brightest pixel
        return [int(value) for value in brightest_pixel]  
    else:
        print(f"Could not read video: {video_path}")  # notifies if the video is unaccessible 
        return None
    cap.release()

def update_csv_with_brightest_color(df, video_folder):
    brightest_colors = []

    for video_id in df['video_id']:
        video_path = f"{video_folder}/{video_id}.mp4"  # reads the videos with the ids in the updated df
        rgb = extract_brightest_color(video_path) # runs the function to extract the brightest colour of the frame
        brightest_colors.append(rgb)

    df['shape_rgb_values'] = brightest_colors # creates another column in the df and stores the value of brightest colours
    return df

video_folder = "./Train_Videos"  # folder with the videos


df= update_csv_with_brightest_color(df, video_folder)
print("done")


df[['R', 'G', 'B']] = pd.DataFrame(df['shape_rgb_values'].tolist(), index=df.index) # the tuple is seperated to different columns as R G B
df

# power: LogisticRegression

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, LabelEncoder
le = LabelEncoder()
df["power"] = le.fit_transform(df["power"])
X = df[['R', 'G', 'B']]  # defines the X value of the model
y = df["power"]             # defines the y value of the model


X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42) # Split data into training and test sets 

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)
X_scaled = scaler.fit_transform(X)
# the values are preprocessed with standardscalar before training with Logistic regression for more accuracy


from sklearn.linear_model import LogisticRegression
model = LogisticRegression()

model.fit(X_train_scaled, y_train)

model.score(X_test_scaled, y_test)



import joblib
joblib.dump(model, './power_model.pkl')
joblib.dump(le, 'label_encoder_power.pkl')
# the model and the label encodes are saved
