==> 5. speed.ipynb <==
# speed: XGBClassifier

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
train_video_features= np.load(r"C:\Users\indra\Desktop\DSG PROJECT\video_features_noiseless_grayscale.npy")
# the train csv is loaded
df= pd.read_csv(r'C:\Users\indra\Desktop\DSG PROJECT\BH25\Training_Data\train.csv')

# since there are only 2 distinct values of speed the values are label encoded as 0 and 1
le = LabelEncoder()
df["speed"] = le.fit_transform(df["speed"])

df

from sklearn.model_selection import train_test_split
label= df["speed"] # label is defined for y value of the model
X_train, X_val, y_train, y_val = train_test_split(train_video_features, label, test_size=0.2, random_state=42)


# XGBClassifier is used for this model
from xgboost import XGBClassifier
model = XGBClassifier(
    n_estimators=200,         
    learning_rate=0.1,        
    max_depth=6,              
    subsample=0.8,            
    colsample_bytree=0.8,     
    random_state=42,          
    use_label_encoder=False,  
    eval_metric="logloss"     
)
# these parameters are derived by hyper parameter tunning with gridsearch cv in other notebook

model.fit(train_video_features, label)

print("done") 

model.score(X_val, y_val) # checked the eval accuracy on seen data

import joblib
joblib.dump(model, './speed_model.pkl')

# the model is saved for further use

joblib.dump(le, 'label_encoder_speed.pkl')
# the label encodes are used for further use
