==> 6. summary.ipynb <==
# summary: RandomForestRegressor

import pandas as pd
import numpy as np
video_features= np.load(r"C:\Users\indra\Desktop\DSG PROJECT\video_features_noiseless_grayscale.npy")
df = pd.read_csv(r'C:\Users\indra\Desktop\DSG PROJECT\BH25\Training_Data\train.csv')


# seperate the values into 2 columns
df[['x_sum', 'y_sum']] = df['video_summary'].str.extract(r'\(([^,]+), ([^,]+)\)')

# Convert columns to numeric values
df['x_sum'] = pd.to_numeric(df['x_sum'])
df['y_sum'] = pd.to_numeric(df['y_sum'])


df

X=video_features # x value of the model is defined
label_x=df["x_sum"] 
label_y=df["y_sum"] # y value of the model is defined for 2 different models

from sklearn.model_selection import train_test_split
X_train_x, X_test_x, y_train_x, y_test_x = train_test_split(X, label_x, test_size=0.2, random_state=42)
X_train_y, X_test_y, y_train_y, y_test_y = train_test_split(X, label_y, test_size=0.2, random_state=42)

# randomforestregressor is used since the values are continuous
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
rf_x = RandomForestRegressor(n_estimators=100, random_state=42) # model is defined

rf_x.fit(X, label_x)

y_pred_x = rf_x.predict(X_test_x)

mse = mean_squared_error(y_test_x, y_pred_x) # MSE is calculated for evaluation

print(f"Mean Squared Error (MSE): {mse:.4f}") 

rf_y = RandomForestRegressor(n_estimators=100, random_state=42) # model is defined

rf_y.fit(X, label_y)

y_pred_y = rf_y.predict(X_test_y)

mse = mean_squared_error(y_test_y, y_pred_y) # MSE is calculate

print(f"Mean Squared Error (MSE): {mse:.4f}")

import joblib
joblib.dump(rf_x, './summary_x_model.pkl')
joblib.dump(rf_y, './summary_y_model.pkl')

# the models are saved for further use
